+++ src/OpenSimBindings/ActionFunctions.cpp
@@ -1692,3 +1692,42 @@ bool osc::ActionTranslatePathPointAndSave(
         return false;  // edit wasn't made
     }
 }
+
+bool osc::ActionTransformPof(
+    UndoableModelStatePair& model,
+    OpenSim::PhysicalOffsetFrame const& pof,
+    glm::vec3 const& deltaPosition,
+    glm::vec3 const& newEulers)
+{
+    OpenSim::ComponentPath const pofPath = osc::GetAbsolutePath(pof);
+    UID const oldVersion = model.getModelVersion();
+    try
+    {
+        OpenSim::Model& mutModel = model.updModel();
+
+        OpenSim::PhysicalOffsetFrame* mutPof = FindComponentMut<OpenSim::PhysicalOffsetFrame>(mutModel, pofPath);
+        if (!mutPof)
+        {
+            model.setModelVersion(oldVersion);  // the provided path isn't a station
+            return false;
+        }
+
+        SimTK::Vec3 const originalPos = mutPof->get_translation();
+        SimTK::Vec3 const newPos = originalPos + ToSimTKVec3(deltaPosition);
+
+        // perform mutation
+        mutPof->set_translation(newPos);
+        mutPof->set_orientation(ToSimTKVec3(newEulers));
+        osc::InitializeModel(mutModel);
+        osc::InitializeState(mutModel);
+
+        return true;
+    }
+    catch (std::exception const& ex)
+    {
+        log::error("error detected while trying to transform a POF: %s", ex.what());
+        model.rollback();
+        return false;
+    }
+    return false;
+}